SystemUI: Allow Wi-Fi/cell tiles to co-exist with provider model
The dedicated Wi-Fi/cellular data QS tiles are partially working now
that we've exposed them in the list of tile options, but they're very
buggy and somewhat broken when the provider model (unified internet) is
enabled.

Allow the tiles to co-exist with InternetTile and the provider model
(including both settings_provider_model and combined signal icons) by
always including QS icons in Wi-Fi/data indicator callbacks, and add
a new flag to indicate whether they should be shown in the internet
tile.

[kdrag0n: ported to 12L. Some conditions could be simplified as both
          cases are now the same, but I've left them as-is to ease
          future porting]

Change-Id: I11f8c3dd4796d792d0f0e48dc21eb5d8158d0476
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>